generalized throttle and added to scrollFire
