+++ jade/page-contents/scrollfire_content.html
@@ -47,53 +47,54 @@
             </tr>
           </tbody>
         </table>
+
+        <!-- Demo -->
+        <h2 class="header">ScrollFire Demo</h2>
+        <p>Scroll through slowly to get sense of what ScrollFire can do for you. This is the ScrollFire code that we have used on this page.</p>
+        <pre>
+    <code class="language-javascript">
+    var options = [</code>
+      <code id="call-1" class="language-javascript">{selector: '#staggered-test', offset: 50, callback: function(el) {
+        Materialize.toast("This is our ScrollFire Demo!", 1500 );
+      } },</code>
+      <code id="call-2" class="language-javascript">{selector: '#staggered-test', offset: 205, callback: function(el) {
+        Materialize.toast("Please continue scrolling!", 1500 );
+      } },</code>
+      <code id="call-3" class="language-javascript">{selector: '#staggered-test', offset: 400, callback: function(el) {
+        Materialize.showStaggeredList($(el));
+      } },</code>
+      <code id="call-4" class="language-javascript">{selector: '#image-test', offset: 500, callback: function(el) {
+        Materialize.fadeInImage($(el));
+      } }</code>
+    ];
+    Materialize.scrollFire(options);
+        </code></pre>
+        <ul id="staggered-test">
+          <li>
+            <h4><a href="#">List Item</a></h4>
+            <p>This is a description</p>
+          </li>
+          <li>
+            <h4><a href="#">List Item</a></h4>
+            <p>This is a description</p>
+          </li>
+          <li>
+            <h4><a href="#">List Item</a></h4>
+            <p>This is a description</p>
+          </li>
+          <li>
+            <h4><a href="#">List Item</a></h4>
+            <p>This is a description</p>
+          </li>
+          <li>
+            <h4><a href="#">List Item</a></h4>
+            <p>This is a description</p>
+          </li>
+        </ul>
+
+        <img id="image-test" class="responsive-img" src="https://unsplash.imgix.net/reserve/nE6neNVdRPSIasnmePZe_IMG_1950.jpg?dpr=1.25&fit=crop&fm=jpg&h=700&q=75&w=1050">
       </div>
 
-      <!-- Demo -->
-      <h2 class="header">ScrollFire Demo</h2>
-      <p>Scroll through slowly to get sense of what ScrollFire can do for you. This is the ScrollFire code that we have used on this page.</p>
-      <pre>
-  <code class="language-javascript">
-  var options = [</code>
-    <code id="call-1" class="language-javascript">{selector: '#staggered-test', offset: 50, callback: function(el) {
-      Materialize.toast("This is our ScrollFire Demo!", 1500 );
-    } },</code>
-    <code id="call-2" class="language-javascript">{selector: '#staggered-test', offset: 205, callback: function(el) {
-      Materialize.toast("Please continue scrolling!", 1500 );
-    } },</code>
-    <code id="call-3" class="language-javascript">{selector: '#staggered-test', offset: 400, callback: function(el) {
-      Materialize.showStaggeredList($(el));
-    } },</code>
-    <code id="call-4" class="language-javascript">{selector: '#image-test', offset: 500, callback: function(el) {
-      Materialize.fadeInImage($(el));
-    } }</code>
-  ];
-  Materialize.scrollFire(options);
-      </code></pre>
-      <ul id="staggered-test">
-        <li>
-          <h4><a href="#">List Item</a></h4>
-          <p>This is a description</p>
-        </li>
-        <li>
-          <h4><a href="#">List Item</a></h4>
-          <p>This is a description</p>
-        </li>
-        <li>
-          <h4><a href="#">List Item</a></h4>
-          <p>This is a description</p>
-        </li>
-        <li>
-          <h4><a href="#">List Item</a></h4>
-          <p>This is a description</p>
-        </li>
-        <li>
-          <h4><a href="#">List Item</a></h4>
-          <p>This is a description</p>
-        </li>
-      </ul>
-
-      <img id="image-test" class="responsive-img" src="https://unsplash.imgix.net/reserve/nE6neNVdRPSIasnmePZe_IMG_1950.jpg?dpr=1.25&fit=crop&fm=jpg&h=700&q=75&w=1050">
 
 
     </div>
